requirements.txt
streamlit
python-dotenv
pandas
pillow
openpyxl

langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-text-splitters

sentence-transformers
faiss-cpu
torch

groq
